updated the phase field section
made some fixes and modifications to the phase field section for the
elements and the materials. Now it seems to work with the nacre files,
need to check if it works with the phase-field paper too.

the × is defined and implemented bu it is not used yet, should check if
it improves the speed